chore: moved app file to src folder
